refactor(WebAPI): Upgraded the unit test system to also use a mock filesystem
