--- src/FileSystem.Contracts/FileSystem/MountOptions.cs
@@ -1,13 +1,0 @@
-﻿namespace FileSystem.Contracts;
-
-public class MountOptions
-{
-    private readonly Dictionary<string, string> _options;
-
-    public MountOptions(Dictionary<string, string> options)
-    {
-        _options = options;
-    }
-
-    public bool IsReadOnly => _options.ContainsKey("ro");
-}
